+++ src/components/UserDetailSections/ReadingRoomAccess/constant.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line import/prefer-default-export
+export const rraColumns = {
+  ACCESS: 'access',
+  READING_ROOM_NAME: 'readingRoomName',
+  NOTES: 'notes',
+  ID: 'id'
+};
+
+export const DEFAULT_SORT_ORDER = rraColumns.ACCESS;